// src/proto/FetchLoginDevice.ts
export namespace Oreka {
    export class FetchLoginDevice extends pb_1.Message {
        #one_of_decls: number[][] = [];
        constructor(data?: any[] | {
            type?: string;
            startTime?: number;
            endTime?: number;
        }) {
            super();
            pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [], this.#one_of_decls);
            if (!Array.isArray(data) && typeof data == "object") {
                if ("type" in data && data.type != undefined) {
                    this.type = data.type;
                }
                if ("startTime" in data && data.startTime != undefined) {
                    this.startTime = data.startTime;
                }
                if ("endTime" in data && data.endTime != undefined) {
                    this.endTime = data.endTime;
                }
            }
        }
        get type() {
            return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
        }
        set type(value: string) {
            pb_1.Message.setField(this, 1, value);
        }
        get startTime() {
            return pb_1.Message.getFieldWithDefault(this, 2, 0) as number;
        }
        set startTime(value: number) {
            pb_1.Message.setField(this, 2, value);
        }
        get endTime() {
            return pb_1.Message.getFieldWithDefault(this, 3, 0) as number;
        }
        set endTime(value: number) {
            pb_1.Message.setField(this, 3, value);
        }
        static fromObject(data: {
            type?: string;
            startTime?: number;
            endTime?: number;
        }): FetchLoginDevice {
            const message = new FetchLoginDevice({});
            if (data.type != null) {
                message.type = data.type;
            }
            if (data.startTime != null) {
                message.startTime = data.startTime;
            }
            if (data.endTime != null) {
                message.endTime = data.endTime;
            }
            return message;
        }
        toObject() {
            const data: {
                type?: string;
                startTime?: number;
                endTime?: number;
            } = {};
            if (this.type != null) {
                data.type = this.type;
            }
            if (this.startTime != null) {
                data.startTime = this.startTime;
            }
            if (this.endTime != null) {
                data.endTime = this.endTime;
            }
            return data;
        }
        serialize(): Uint8Array;
        serialize(w: pb_1.BinaryWriter): void;
        serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
            const writer = w || new pb_1.BinaryWriter();
            if (this.type.length)
                writer.writeString(1, this.type);
            if (this.startTime != 0)
                writer.writeUint32(2, this.startTime);
            if (this.endTime != 0)
                writer.writeUint32(3, this.endTime);
            if (!w)
                return writer.getResultBuffer();
        }
        static deserialize(bytes: Uint8Array | pb_1.BinaryReader): FetchLoginDevice {
            const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new FetchLoginDevice();
            while (reader.nextField()) {
                if (reader.isEndGroup())
                    break;
                switch (reader.getFieldNumber()) {
                    case 1:
                        message.type = reader.readString();
                        break;
                    case 2:
                        message.startTime = reader.readUint32();
                        break;
                    case 3:
                        message.endTime = reader.readUint32();
                        break;
                    default: reader.skipField();
                }
            }
            return message;
        }
        serializeBinary(): Uint8Array {
            return this.serialize();
        }
        static deserializeBinary(bytes: Uint8Array): FetchLoginDevice {
            return FetchLoginDevice.deserialize(bytes);
        }
    }
}